requirements.txt
python-memcached==1.59
Flask==1.1.2
prometheus-client==0.8.0